review_agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cxxtract-review-agent"
version = "0.1.0"
description = "Standalone semantic C++ PR review agent powered by CXXtract2."
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.6.0",
    "pyyaml>=6.0.1",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
review-agent = "review_agent.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["review_agent"]

